+++ src/library/components/table/index.tsx
@@ -1,6 +1,6 @@
 import React, { PropsWithChildren, useState, useEffect, useRef, useCallback, useMemo } from 'react';
 import ReactDOM from 'react-dom';
-import { DataTableProperties, ColumnVisibilityStorage, DataFnResult, ColumnSorts, QSColumnSorts, QueryFilterGroup, EditFormData, QuickEditFormData, QSGroupBy, GroupBy, ColumnSort, ColumnConfigurationWithGroup } from './types';
+import { DataTableProperties, ColumnVisibilityStorage, DataFnResult, ColumnSorts, QSColumnSorts, QueryFilterGroup, EditFormData, QuickEditFormData, QSGroupBy, GroupBy, ColumnSort, ColumnConfigurationWithGroup, Pagination } from './types';
 import { useDeepDerivedState } from '../../utils/useDerivedState';
 import { useQueryState, batchedQSUpdate } from '@borvik/use-querystate';
 import { transformColumns, getFlattenedColumns, generateHeaderRows } from '../../utils/transformColumnProps';
@@ -20,12 +20,15 @@ import { TableEditorButton } from './editors/editButton';
 import { TableActionButtons } from './actions';
 import { getRowKey } from '../../utils/getRowKey';
 import { update } from '../../utils/immutable';
+import { QueryString } from '@borvik/querystring';
+import { defaults } from 'lodash';
+import { DeepPartial } from '@borvik/use-querystate/dist/types';
 
 const primaryKeyWarned: {[x:string]: boolean} = {};
 const fixedLeftWarned: Record<string, boolean> = {};
 const fixedRightWarned: Record<string, boolean> = {};
 
-export const DataTable = function DataTable<T>({paginate = 'both', quickEditPosition = 'both', hideSearchForm = false, ...props}: PropsWithChildren<DataTableProperties<T>>) {
+export const DataTable = function DataTable<T>({paginate = 'both', quickEditPosition = 'both', hideSearchForm = false, defaultFilter, ...props}: PropsWithChildren<DataTableProperties<T>>) {
   const canGroupBy = !!props.canGroupBy && !!props.multiColumnSorts;
 
   /**
@@ -50,8 +53,8 @@ export const DataTable = function DataTable<T>({paginate = 'both', quickEditPosi
   const [qsGroupBy, setGroupBy] = useParsedQs<GroupBy, QSGroupBy>(
     { group: props.defaultGroupBy ?? [] },
     (qsSort) => ({ // parse
-      group: qsSort.group.map(v => {
-        let parts = v.split(' ').filter(a => !!a);
+      group: (qsSort?.group ?? []).map(v => {
+        let parts = v!.split(' ').filter(a => !!a);
         if (parts.length !== 2) return null;
         return {
           column: parts[0],
@@ -169,15 +172,45 @@ export const DataTable = function DataTable<T>({paginate = 'both', quickEditPosi
   const [editFormData, setFormData] = useState<EditFormData>({});
   const [selectedRows, setSelectedRows] = useState<Record<string | number, T>>({});
 
-  const [pagination, setPagination] = useQueryState({page: 1, perPage: props.paginateOptions?.defaultPerPage ?? 10}, {
-    ...props.qs
-  });
+  const [pagination, setPagination] = useParsedQs<Pagination, DeepPartial<Pagination>>(
+    {page: 1, perPage: props.paginateOptions?.defaultPerPage ?? 10},
+    (qsPagination) => ({
+      page: qsPagination.page ?? 1,
+      perPage: qsPagination.perPage ?? props.paginateOptions?.defaultPerPage ?? 10,
+    }),
+    (state) => state,
+    {
+      ...props.qs
+    }
+  )
 
   const [searchQuery, setSearchQuery] = useQueryState({query: ''}, {
     ...props.qs
   });
 
-  const [rawFilter, setRawFilter] = useQueryState<{filter?: any}>({}, {
+  const { defaultRawFilter, defaultConvertedFilter } = useMemo(() => {
+    if (typeof defaultFilter === 'string' || typeof defaultFilter === 'undefined') {
+      if (!defaultFilter?.trim()) {
+        return { defaultRawFilter: {}, defaultConvertedFilter: { groupOperator: 'and', filters: [] } as QueryFilterGroup };
+      }
+
+      // parse the filter string like query
+      let parsedDefaultRawFilter = QueryString.parse(defaultFilter);
+      parsedDefaultRawFilter = Object.keys(parsedDefaultRawFilter).length ? { filter: parsedDefaultRawFilter } : {};
+      return {
+        defaultRawFilter: parsedDefaultRawFilter,
+        defaultConvertedFilter: convertFromQS(parsedDefaultRawFilter, filterColumns),
+      };
+    }
+
+    const defaultRawFilter = { filter: defaultFilter };
+    return {
+      defaultRawFilter,
+      defaultConvertedFilter: convertFromQS(defaultRawFilter, filterColumns),
+    };
+  }, [ defaultFilter, filterColumns ]);
+
+  const [rawFilter, setRawFilter] = useQueryState<{filter?: any}>(defaultRawFilter, {
     ...props.qs,
     types: {
       filter: 'any'
@@ -186,7 +219,7 @@ export const DataTable = function DataTable<T>({paginate = 'both', quickEditPosi
   });
 
   const [filter, setFilter] = useParsedQs<QueryFilterGroup, {filter?: any}>(
-    { groupOperator: 'and', filters: [] },
+    defaultConvertedFilter,
     (qsFilter) => convertFromQS(qsFilter, filterColumns),
     (state) => convertToQS(state, filterColumns),
     {
@@ -201,8 +234,8 @@ export const DataTable = function DataTable<T>({paginate = 'both', quickEditPosi
   const [columnSort, setColumnSort] = useParsedQs<ColumnSorts, QSColumnSorts>(
     { sort: props.defaultSort ?? [] },
     (qsSort) => ({ // parse
-      sort: qsSort.sort.map(v => {
-        let parts = v.split(' ').filter(a => !!a);
+      sort: (qsSort?.sort ?? []).map(v => {
+        let parts = v!.split(' ').filter(a => !!a);
         if (parts.length !== 2) return null;
         return {
           column: parts[0],
@@ -249,7 +282,7 @@ export const DataTable = function DataTable<T>({paginate = 'both', quickEditPosi
       if (typeof props.onQueryChange === 'function') {
         props.onQueryChange({
           pagination,
-          search: hideSearchForm ? '' : searchQuery.query,
+          search: hideSearchForm ? '' : (searchQuery.query ?? ''),
           sorts: fullSort,
           filters: filter.filters.length ? filter : undefined,
         });
@@ -258,7 +291,7 @@ export const DataTable = function DataTable<T>({paginate = 'both', quickEditPosi
       if (typeof props.data === 'function') {
         let returnedData = await props.data({
           pagination,
-          search: hideSearchForm ? '' : searchQuery.query,
+          search: hideSearchForm ? '' : (searchQuery.query ?? ''),
           sorts: fullSort,
           filters: filter.filters.length ? filter : undefined,
         }, doSetDataList);
@@ -407,14 +440,14 @@ export const DataTable = function DataTable<T>({paginate = 'both', quickEditPosi
   const searchFormOnSearch = useCallback((query: string) => {
     batchedQSUpdate(() => {
       setSearchQuery({ query });
-      setPagination({ page: 1 });
+      setPagination(prev => ({ page: 1, perPage: prev.perPage }));
     });
   }, [setSearchQuery, setPagination]);
 
   const searchFormApplyFilter = useCallback((newFilter: any) => {
     batchedQSUpdate(() => {
       setRawFilter({ filter: newFilter });
-      setPagination({ page: 1 });
+      setPagination(prev => ({ page: 1, perPage: prev.perPage }));
     });
   }, [setRawFilter, setPagination]);
 
@@ -473,7 +506,7 @@ export const DataTable = function DataTable<T>({paginate = 'both', quickEditPosi
         <div ref={topEl} className={`ts-datatable-top`}>
           <div className='ts-datatable-search-filters'>
             {!hideSearchForm && <SearchForm
-              searchQuery={searchQuery.query}
+              searchQuery={searchQuery.query ?? ''}
               filter={rawFilter.filter}
               onSearch={searchFormOnSearch}
               applyFilter={searchFormApplyFilter}
@@ -499,7 +532,7 @@ export const DataTable = function DataTable<T>({paginate = 'both', quickEditPosi
               <Paginate
                 {...props.paginateOptions}
                 {...pagination}
-                changePage={(page) => setPagination(page)}
+                changePage={(page) => setPagination(prev => ({ ...prev, ...page }))}
                 total={typeof props.data === 'function' ? stateDataList.total : props.totalCount}
             />}
           </div>
@@ -524,7 +557,7 @@ export const DataTable = function DataTable<T>({paginate = 'both', quickEditPosi
             <Paginate
               {...props.paginateOptions}
               {...pagination}
-              changePage={(page) => setPagination(page)}
+              changePage={(page) => setPagination(prev => ({ ...prev, ...page }))}
               total={typeof props.data === 'function' ? stateDataList.total : props.totalCount}
             />
           </div>}