+++ src/library/utils/useParsedQS.ts
@@ -1,18 +1,18 @@
 import { useQueryState } from '@borvik/use-querystate';
-import { QueryStateOptions } from '@borvik/use-querystate/dist/types';
+import { QueryStateOptions, DeepPartial } from '@borvik/use-querystate/dist/types';
 import { useDeepDerivedState } from "./useDerivedState";
 import { useCallback, useDebugValue } from "react";
 
 type ConvertFn<T, V> = (value: T) => V;
 
-export function useParsedQs<State, QSState extends object>(initialState: State, parse: ConvertFn<QSState, State>, encode: ConvertFn<State, QSState>, options?: QueryStateOptions): [State, (newState: State | ((state: State) => State)) => void] {
+export function useParsedQs<State, QSState extends object>(initialState: State, parse: ConvertFn<DeepPartial<QSState>, State>, encode: ConvertFn<State, DeepPartial<QSState>>, options?: QueryStateOptions): [State, (newState: State | ((state: State) => State)) => void] {
   const [qsInitialState] = useDeepDerivedState(() => encode(initialState), [ initialState ]);
-  const [localState, setLocalState] = useQueryState(qsInitialState, options);
+  const [localState, setLocalState] = useQueryState<QSState>(qsInitialState as any, options);
   const [actualState] = useDeepDerivedState(() => parse(localState), [localState]);
 
   const publicSetState = useCallback(
     (newState: (State | ((state: State) => State))) => {
-      setLocalState(ls => {
+      setLocalState(() => {
         let newActualState = typeof newState === 'function'
           ? (newState as any)(actualState)
           : newState;